Fixed warning

+++ source/newtonian/two_dimensional/simple_cell_updater.cpp
@@ -43,7 +43,7 @@ namespace
 		Vector2D Vcell = cell.momentum / cell.mass;
 		double Et = cell.energy / cell.mass - 0.5*ScalarProd(Vcell, Vcell);
 		for (size_t i = 0; i < N; ++i)
-			if (neigh[i] < static_cast<int>(Norg) || tess.GetOriginalIndex(index) == static_cast<int>(index))
+			if (neigh[i] < static_cast<int>(Norg) || tess.GetOriginalIndex(static_cast<int>(index)) == static_cast<int>(index))
 				maxDV = std::max(maxDV, abs(Vcell - cells.at(neigh[i]).momentum / cells[neigh[i]].mass));
 		return 0.005*maxDV*maxDV > Et;
 	}
